requirements.txt
openai-whisper
flask
flask-cors
requests
pydub
opencc-python-reimplemented
httpx